feat: create users, item router for readability
