+++ .pre-commit-config.yaml
@@ -9,11 +9,11 @@ repos:
     #   rev: 22.10.0
     #   hooks:
     #       - id: black
-    # - repo: https://github.com/pycqa/isort
-    #   rev: 5.12.0
-    #   hooks:
-    #       - id: isort
-    #         name: isort (python)
+    - repo: https://github.com/pycqa/isort
+      rev: 5.12.0
+      hooks:
+          - id: isort
+            name: isort (python)
     # - repo: https://github.com/charliermarsh/ruff-pre-commit
     #   rev: "v0.0.265"
     #   hooks:
