+++ client/src/components/Login/login.jsx
@@ -96,7 +96,7 @@ function Login({ closeModal = () => {} }) {
 
     const handleSignup = () => {
         console.log(`signup => ID: ${id}, Password: ${password}, ConfirmPassword: ${confirm}`);
-        
+
         const SignUpParams = {
             user_name: id,
             user_pwd: password,
