+++ scripts/insert_guest.py
@@ -1,17 +1,19 @@
 import psycopg2
 
-host = 'localhost'
-port = '6000'
-database = 'postgres'
-user = 'postgres'
-password = 'password'
+host = "localhost"
+port = "6000"
+database = "postgres"
+user = "postgres"
+password = "password"
 
-conn = psycopg2.connect(host=host, port=port, database=database, user=user, password=password)
+conn = psycopg2.connect(
+    host=host, port=port, database=database, user=user, password=password
+)
 cursor = conn.cursor()
 
 # 삽입할 데이터
-user_name = 'guest'
-user_pwd = 'guest_pwd'
+user_name = "guest"
+user_pwd = "guest_pwd"
 
 # INSERT 쿼리 실행
 query = 'INSERT INTO "user" (user_name, user_pwd) VALUES (%s, %s)'
