+++ scripts/insert_outfit.py
@@ -1,23 +1,26 @@
-import psycopg2
-import csv
 import ast
+import csv
 
-host = 'localhost'
-port = '6000'
-database = 'postgres'
-user = 'postgres'
-password = 'password'
+import psycopg2
 
-conn = psycopg2.connect(host=host, port=port, database=database, user=user, password=password)
+host = "localhost"
+port = "6000"
+database = "postgres"
+user = "postgres"
+password = "password"
 
-csv_file = '../meta_22-23.csv'
+conn = psycopg2.connect(
+    host=host, port=port, database=database, user=user, password=password
+)
+
+csv_file = "../meta_22-23.csv"
 
 cursor = conn.cursor()
-delete_query = 'DELETE FROM outfit'
+delete_query = "DELETE FROM outfit"
 cursor.execute(delete_query)
 conn.commit()
 
-with open(csv_file, 'r') as f:
+with open(csv_file, "r") as f:
     reader = csv.reader(f)
     next(reader)
 
@@ -26,7 +29,7 @@ with open(csv_file, 'r') as f:
         outfit_id = int(row[0])
 
         # outfit_id가 이미 존재하는지 확인
-        check_query = 'SELECT COUNT(*) FROM outfit WHERE outfit_id = %s'
+        check_query = "SELECT COUNT(*) FROM outfit WHERE outfit_id = %s"
         cursor.execute(check_query, (outfit_id,))
         count = cursor.fetchone()[0]
 
@@ -34,10 +37,10 @@ with open(csv_file, 'r') as f:
             print(f"Skipping duplicate outfit_id: {outfit_id}")
             continue
 
-        age = int(row[2]) if row[2] != '연령미상' else None
-        occupation = row[9] if row[9] != '정보없음' else None
+        age = int(row[2]) if row[2] != "연령미상" else None
+        occupation = row[9] if row[9] != "정보없음" else None
         tags = ast.literal_eval(row[6])
-        brands = ast.literal_eval(row[7]) if row[7] != '[]' else None
+        brands = ast.literal_eval(row[7]) if row[7] != "[]" else None
 
         query = 'INSERT INTO outfit (outfit_id, gender, age, img_url, origin_url, reporter, tags, brands, region, occupation, style, "date") VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s)'
         values = (
@@ -52,7 +55,7 @@ with open(csv_file, 'r') as f:
             row[8],  # region: VARCHAR
             occupation,  # occupation: VARCHAR
             row[10],  # style: VARCHAR
-            row[11]  # date: TIMESTAMP WITHOUT TIME ZONE
+            row[11],  # date: TIMESTAMP WITHOUT TIME ZONE
         )
         cursor.execute(query, values)
