+++ scripts/insert_similar.py
@@ -1,25 +1,30 @@
 import csv
+
 import psycopg2
 
-host = 'localhost'
-port = '6000'
-database = 'postgres'
-user = 'postgres'
-password = 'password'
+host = "localhost"
+port = "6000"
+database = "postgres"
+user = "postgres"
+password = "password"
 
-csv_file = '../top3_22-23.csv'
+csv_file = "../top3_22-23.csv"
 
-conn = psycopg2.connect(host=host, port=port, database=database, user=user, password=password)
+conn = psycopg2.connect(
+    host=host, port=port, database=database, user=user, password=password
+)
 cursor = conn.cursor()
 
 delete_query = 'DELETE FROM "similar"'
 cursor.execute(delete_query)
 
-with open(csv_file, 'r') as f:
+with open(csv_file, "r") as f:
     reader = csv.DictReader(f)
     for row in reader:
-        outfit_id = int(row['id'])
-        similar_outfits = [int(outfit.strip()) for outfit in row['top3'].strip('[]').split(',')]
+        outfit_id = int(row["id"])
+        similar_outfits = [
+            int(outfit.strip()) for outfit in row["top3"].strip("[]").split(",")
+        ]
 
         # outfit_id가 "outfit" 테이블에 존재하는지 확인
         check_query = 'SELECT COUNT(*) FROM "outfit" WHERE outfit_id = %s'
@@ -30,7 +35,9 @@ with open(csv_file, 'r') as f:
             print(f"Skipping outfit_id not found in 'outfit' table: {outfit_id}")
             continue
 
-        similar_outfits_str = "{" + ",".join(str(outfit) for outfit in similar_outfits) + "}"
+        similar_outfits_str = (
+            "{" + ",".join(str(outfit) for outfit in similar_outfits) + "}"
+        )
 
         # outfit_id가 이미 존재하는지 확인
         check_query = 'SELECT COUNT(*) FROM "similar" WHERE outfit_id = %s'
