+++ src/database.py
@@ -1,6 +1,6 @@
-from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy import create_engine
-from sqlalchemy.orm import Session, sessionmaker
+from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy.orm import sessionmaker
 
 SQLALCHEMY_DATABASE_URL = "postgresql://postgres:password@postgres/postgres"
 engine = create_engine(SQLALCHEMY_DATABASE_URL, echo=False)
